FIXED the CustomerSearch Panel
Fixed the Customer search panel.
Removed the unused prop from CustomerSearch that was not the issue.

The issue was, because of the customerPanelPage now using stores for fetching data, I had the id of the store as customer exactly the same as "searchCustomerIbanByNameStore". So just by changing the name in customerProfileStore, I fixed the page breaking issue.